django_backend/requirement.txt
amqp==5.2.0
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.2.0
filelock==3.15.4
flower==2.0.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.5
humanize==4.10.0
idna==3.7
jiter==0.5.0
kombu==5.3.7
openai==1.35.10
packaging==24.1
peopledatalabs==3.1.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
